Create panel manager for explore page
