+++ app/javascript/styles/_explore.scss
@@ -153,7 +153,6 @@
 
 .view-options-toggle-on {
   position: absolute;
-  top: 0px;
   right: 1px;
   padding: 0.5em;
   background: #fff;
